requirements.txt
attrs
keras
numpy==1.21
tensorflow
